+++ federation/hostmeta/django/generators.py
@@ -5,7 +5,7 @@ from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.http import HttpResponseBadRequest, JsonResponse, HttpResponseNotFound
 
-from federation.hostmeta.generators import RFC3033Webfinger
+from federation.hostmeta.generators import RFC3033Webfinger, generate_nodeinfo2_document
 
 logger = logging.getLogger("federation")
 
@@ -16,6 +16,7 @@ def get_configuration():
     """
     configuration = {
         "hcard_path": "/hcard/users/",
+        "nodeinfo2_function": None,
         "search_path": None,
     }
     configuration.update(settings.FEDERATION)
@@ -27,16 +28,26 @@ def get_configuration():
     return configuration
 
 
-def get_profile_func():
+def get_function_from_config(item):
     """
     Import the function to get profile by handle.
     """
     config = get_configuration()
-    profile_func_path = config.get("get_profile_function")
-    module_path, func_name = profile_func_path.rsplit(".", 1)
+    func_path = config.get(item)
+    module_path, func_name = func_path.rsplit(".", 1)
     module = importlib.import_module(module_path)
-    profile_func = getattr(module, func_name)
-    return profile_func
+    func = getattr(module, func_name)
+    return func
+
+
+def nodeinfo2_view(request, *args, **kwargs):
+    try:
+        nodeinfo2_func = get_function_from_config("nodeinfo2_function")
+    except AttributeError:
+        return HttpResponseBadRequest("Not configured")
+    nodeinfo2 = nodeinfo2_func()
+
+    return JsonResponse(generate_nodeinfo2_document(**nodeinfo2))
 
 
 def rfc3033_webfinger_view(request, *args, **kwargs):
@@ -50,7 +61,7 @@ def rfc3033_webfinger_view(request, *args, **kwargs):
         return HttpResponseBadRequest("Invalid resource")
 
     handle = resource.replace("acct:", "")
-    profile_func = get_profile_func()
+    profile_func = get_function_from_config("get_profile_function")
 
     try:
         profile = profile_func(handle)